KU-88/refactor: Event/Party 관련 타입 정의를 types 파일로 분리
