+++ mocks/parties.ts
@@ -1,40 +1,6 @@
-export type PartyStatus = 'RECRUITING' | 'FULL' | 'CLOSED';
-export type PartyType = 'SCHEDULE_AND_TRANSPORT' | 'SCHEDULE_ONLY' | 'TRANSPORT_ONLY';
-export type GenderType = 'MALE' | 'FEMALE' | 'ALL';
-export type RouteType = 'ONE_WAY' | 'ROUND_TRIP';
+import type { PartyItem } from "@/types/party"; 
 
-export interface PartyItem {
-  partyId: number;
-  type: PartyType;
-  routeType: RouteType;
-  status: PartyStatus;
-  gender: GenderType;
-  maxBirthYear: string;
-  minBirthYear: string;
-  startAt: string;
-  endAt: string;
-  description: string;
-  chatUrl: string | null;
-  capacity: number;
-  currentParticipants: number;
-    host: {
-    memberId: number;
-    nickname: string | null;
-    profileImageUrl: number;
-    role: 'HOST' | 'PARTICIPANT';
-  };
-  location: {
-    placeName: string;
-    addressName: string;
-    roadAddressName: string;
-    latitude: number;
-    longitude: number;
-    startLocation: string;
-  };
-  isLiked: boolean;
-}
-
-export const events: PartyItem[] = [
+export const parties: PartyItem[] = [
     {
         "partyId": 1,
         "type": "TRANSPORT_ONLY",
